--- a/src/Xamarin.Android.Build.Tasks/Resources/XamarinAndroidEnvironmentVariables.java
+++ b/src/Xamarin.Android.Build.Tasks/Resources/XamarinAndroidEnvironmentVariables.java
@@ -1,0 +1,9 @@
+package mono.android.app;
+
+public class XamarinAndroidEnvironmentVariables
+{
+	// Variables are specified the in "name", "value" pairs
+	public static final String[] Variables = new String[] {
+//@ENVVARS@
+	};
+}