--- a/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Build.Tasks.csproj
+++ b/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Build.Tasks.csproj
@@ -686,6 +686,9 @@
     <EmbeddedResource Include="Resources\ResourcePatcher.java">
       <LogicalName>ResourcePatcher.java</LogicalName>
     </EmbeddedResource>
+    <EmbeddedResource Include="Resources\XamarinAndroidEnvironmentVariables.java">
+      <LogicalName>XamarinAndroidEnvironmentVariables.java</LogicalName>
+    </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Xamarin.Android.Tools.Aidl\Xamarin.Android.Tools.Aidl.csproj">