… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/BuildTest.cs
… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/PackagingTest.cs
… src/monodroid/CMakeLists.txt
… src/monodroid/jni/android-system.cc
… src/monodroid/jni/debug.cc
… src/monodroid/jni/dylib-mono.cc
… src/monodroid/jni/dylib-mono.h
… src/monodroid/jni/embedded-assemblies.cc
… src/monodroid/jni/jni-wrappers.h
… src/monodroid/jni/mono_android_Runtime.h
… src/monodroid/jni/monodroid-glue-internal.h
… src/monodroid/jni/monodroid-glue.cc
… src/monodroid/jni/monodroid.h
… src/monodroid/jni/new_delete.cc
… src/monodroid/jni/osbridge.cc
… src/monodroid/jni/osbridge.h
… src/monodroid/jni/util.cc
… src/monodroid/jni/util.h
… src/monodroid/jni/xamarin_getifaddrs.cc
… src/monodroid/monodroid.targets
… tests/EmbeddedDSOs/EmbeddedDSO-UnitTests/BuildTests.cs
… tests/Xamarin.Forms-Performance-Integration/Droid/Properties/AndroidManifest.xml
… tests/Xamarin.Forms-Performance-Integration/Droid/Xamarin.Forms.Performance.Integration.Droid.csproj