@garuma 2018-12-10
Would it be possible to have an equivalent branch on monodroid? That way I can both buddy test manually and launch a CI run on the changes.