@grendello 2018-12-12
Note: we're currently using a tiny bit of STL stuff (`std::atomic` and `std::mutex`) and I also discovered that clang in the NDK does **not** remove RTTI or exception stuff despite being passed `-fno-rtti` and `-fno-exceptions` (gcc correctly removes everything related to RTTI and exceptions). I think, for now, the use of STL can stay until we have time to replace those features with something local. Regarding clang, I'm going to investigate that.