@grendello 2018-12-12
I need to figure out what's going on with RTTI and exceptions before this can be merged