@grendello 2018-12-12
Stil `do-not-merge` because I haven't implemented `std::mutex` and `std::lock_guard` yet (hopefully later today), but the RTTI and exception stuff is now gone.