@grendello 2018-12-12
`std::mutex` and `std::lock_guard` implemented, still need to implement `std::atomic`