@grendello 2018-12-12
We don't need `std::atomic`