@jonpryor 2019-01-04
Looks like the [PR build hung](https://jenkins.mono-project.com/job/xamarin-android-pr-builder-release/286/console):

```
18:59:35   Results (nunit2) saved as TestResult-Xamarin.Android.Build.Tests.xml
```

Note: at the time of this comment, it is 21:51, so we're approaching the 3h timeout value.

I'm aborting the build and will restart it.

(I wanted to restart the build *anyway*, as I added the **full-mono-integration-build** label...)