@jonpryor 2019-01-04
The most recent [PR build ](
https://jenkins.mono-project.com/job/xamarin-android-pr-builder-release/290/) failed in the MXE portion of `src/monodroid`:

```
  FAILED: : && /Users/builder/android-toolchain/mxe-b9cbb53/bin/x86_64-w64-mingw32.static-g++   ...
  Cannot export _ZN7xamarin7android4Util19package_hash_to_hexIiIEEEvjT_DpT0_: symbol wrong type (4 vs 3)
  Cannot export _ZN7xamarin7android4Util19package_hash_to_hexIiIiEEEvjT_DpT0_: symbol wrong type (4 vs 3)
  Cannot export _ZN7xamarin7android4Util19package_hash_to_hexIiIiiEEEvjT_DpT0_: symbol wrong type (4 vs 3)
  Cannot export _ZN7xamarin7android4Util19package_hash_to_hexIiIiiiEEEvjT_DpT0_: symbol wrong type (4 vs 3)
  Cannot export _ZN7xamarin7android4Util19package_hash_to_hexIiIiiiiEEEvjT_DpT0_: symbol wrong type (4 vs 3)
  Cannot export _ZN7xamarin7android4Util19package_hash_to_hexIiIiiiiiEEEvjT_DpT0_: symbol wrong type (4 vs 3)
  Cannot export _ZN7xamarin7android4Util19package_hash_to_hexIiIiiiiiiEEEvjT_DpT0_: symbol wrong type (4 vs 3)
  Cannot export _ZN7xamarin7android4Util19package_hash_to_hexIiIiiiiiiiEEEvjT_DpT0_: symbol wrong type (4 vs 3)
collect2 : error : ld returned 1 exit status [/Users/builder/jenkins/workspace/xamarin-android-pr-builder-release/xamarin-android/src/monodroid/monodroid.csproj]
```

wat?

It looks like gcc/ld doesn't like the template meta-programming within `package_hash_to_hex()`.

There's also quite a few warnings about signed vs. unsigned integer comparisons:

```
src/monodroid/jni/android-system.cc(592,20): warning G971A9B88: comparison between signed and unsigned integer expressions [-Wsign-compare]
    for (int i = 0; i < AndroidSystem::MAX_OVERRIDES; i++) {
                      ^
```