@grendello 2019-01-04
@jonpryor gcc works just fine with the metaprogramming bit, it's mxe gcc that's too old.